fix: custom tags (<edited>) should never be self-closing
Self-closing tags are only allowed for void elements in HTML5.
Otherwise, the required behavior for parsers is to ignore the slash and
consider the expression as a tag opening. This will lead to bugs with
htmlparser2 if `<edited />` was followed by other tags, where those tags
would be parsed as children of `<edited>` instead of siblings.