Handle the subStep directly from Onyx
The subState in the VBA flow was being copied from Onyx into a local
state in the BankAccountStep component and it was read from this local
state. This caused that when we update the subState from BankAccounts.js
in Onyx, the component BankAccountStep missed the update because
it was reading it from the local state.